Compiles on most platforms, starts. Not really tested.

Conflicts:
	acinclude.m4
	lib/libdvbsub/Makefile.am
	lib/libtriple/Makefile.am
	lib/libtuxtxt/Makefile.am
	src/Makefile.am
	src/create_rcsim_h.sh
	src/daemonc/Makefile.am
	src/driver/Makefile.am
	src/driver/audiodec/Makefile.am
	src/driver/pictureviewer/Makefile.am
	src/eitd/Makefile.am
	src/gui/Makefile.am
	src/gui/bouquetlist.cpp
	src/gui/channellist.cpp
	src/gui/scan_setup.cpp
	src/gui/streaminfo2.cpp
	src/gui/streaminfo2.h
	src/gui/update.cpp
	src/gui/widget/Makefile.am
	src/gui/widget/listbox.cpp
	src/neutrino.cpp
	src/nhttpd/tuxboxapi/coolstream/Makefile.am
	src/rcsim.c
	src/system/Makefile.am
	src/zapit/src/Makefile.am
	src/zapit/src/frontend.cpp
	src/zapit/src/zapit.cpp

+++ src/gui/infoviewer_bb.cpp
@@ -640,9 +640,9 @@ void CInfoViewerBB::showSysfsHdd()
 	if (g_settings.infobar_show_sysfs_hdd) {
 		//sysFS info
 		int percent = 0;
-		long t, u;
+		uint64_t t, u;
 		if (get_fs_usage("/", t, u))
-			percent = (u * 100ULL) / t;
+			percent = (int)((u * 100ULL) / t);
 		showBarSys(percent);
 
 		if (check_dir(g_settings.network_nfs_recordingdir) == 0)
@@ -655,9 +655,9 @@ void CInfoViewerBB::showSysfsHdd()
 void* CInfoViewerBB::hddperThread(void *arg)
 {
 	CInfoViewerBB *infoViewerBB = (CInfoViewerBB*) arg;
-	long t, u;
+	uint64_t t, u;
 	if (get_fs_usage(g_settings.network_nfs_recordingdir, t, u))
-		infoViewerBB->hddpercent = (u * 100ULL) / t;
+		infoViewerBB->hddpercent = (int)((u * 100ULL) / t);
 	else
 		infoViewerBB->hddpercent = 0;
 	infoViewerBB->hddperTflag=false;
